Tentative support for Raspberry PI HW mouse cursor.
Currently not supported by Qemu, so untested.
Still todo: hiding the mouse cursor, which can prove complicated, as with a hardware sprite, we don't need to flash it
every time we draw to the screen, which sadly uses the same mechanism to hide and show the cursor. (This should have
been abstracted by the VDI instead of having the applications take care of it.)